+++ config/EURUSD_M1.h
@@ -24,7 +24,7 @@ struct Stg_ADX_Params_M1 : StgParams {
     signal_close_method = 0;
     signal_close_level = (float)0;
     price_stop_method = 0;
-    price_stop_level = 1;
+    price_stop_level = (float)1;
     tick_filter_method = 1;
     max_spread = 0;
   }
